figured out how to do min and max damage stats

+++ src/main/java/com/robertx22/Stats/FillableStat.java
@@ -2,6 +2,6 @@ package com.robertx22.stats;
 
 public abstract class FillableStat extends Stat {
 
-	public Double CurrentValue;
+	public int CurrentValue;
 
 }
